Improve handling of date types.
pyodbc takes care of converting dates to their appropriate string
representation, so we just need to pass the date/time/datetime instance
to pyodbc and it will take care of the rest.

We need to make sure that we are passing the correct type though, as
passing a datetime instead of a date can cause a Type Error in the
database query.